backend/requirements.txt


# ========== FRAMEWORK WEB ==========
fastapi==0.111.0
uvicorn==0.30.0

# ========== ENVIRONNEMENT ET CONFIG ==========
python-dotenv==1.0.1

# ========== MACHINE LEARNING / IA ==========
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2

# ========== LLM ET RAG ==========
langchain==0.2.3
faiss-cpu==1.8.0
sentence-transformers==2.7.0
gpt4all==2.7.0

# ========== BASE DE DONNÉES ==========
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9    # si tu utilises PostgreSQL
# sqlite fonctionne sans package supplémentaire

# ========== API ET SCHEMAS ==========
pydantic==2.8.2

# ========== FRONTEND / COMMUNICATION ==========
jinja2==3.1.4  # si tu veux faire des templates HTML
requests==2.32.3

# ========== SÉCURITÉ ==========
python-jose==3.3.0   # pour JWT
passlib[bcrypt]==1.7.4

# ========== TESTS ET DEBUG ==========
httpx==0.27.0
pytest==8.3.1



